back-end/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
psycopg[binary]
chromadb
ollama
tqdm
colorama
aiofiles
fastapi-cors